Cryptopals 23
the better to untemper with

wip

mt19937 untemper

not performant - it brute forces the to find the periodicity of the
particular shift/mask step and then returns the previous step.  I left a
function that prints out another algorithm that mostly makes sense but
not enough that I'm comfortable doing it myself.  I *think* I can distil
the two algorithms into a performant general-case, but I'm not 100% yet
and it's late.

better comment

whoops, we want to print the 32 least sig bits

hoist parameters_for

implement ==

spec